+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - The method name from Board.PlacePieces() to SetupPieces().
+- Board.Columns and Board.Rows to readonly.
 
 ## [0.0.2] - 2020-11-02